* 'qt-master-from-4.7' of scm.dev.nokia.troll.no:qt/qt-integration: (241 commits)
  Doc: Fixed a typo
  Doc: QAbstractItemModel: note unexpected behavior with beginMoveRows
  Added environment variable graphicssystem switch.
  Fixed a problem where QPixmaps where re-bound in non-sharing contexts.
  Fix bug in QGraphicsItem::scroll() when called with no QRectF argument.
  Make sure we don't pull inn /usr/X11 stuff in the qws mkspek conf
  Avoid unnecessary detach / deep copy in QGraphicsItem::scroll().
  Missed file.
  Documentation.
  Optimization for sci file loading.
  Ensure text color set by S60 input method works for QML Items
  License.
  Avoid refilling view during model change.
  Test positionAt.
  Fix positionAt when autoScroll is use.
  Fix example code and image
  Fix doc
  Fix some bugs in the QML Web Browser buttons
  Fix zoom bug in QML Web Browser demo
  Fix Typo
  ...

+++ src/declarative/graphicsitems/qdeclarativeanchors_p_p.h
@@ -131,7 +131,7 @@ public:
     bool checkVValid() const;
     bool checkHAnchorValid(QDeclarativeAnchorLine anchor) const;
     bool checkVAnchorValid(QDeclarativeAnchorLine anchor) const;
-    bool calcStretch(const QDeclarativeAnchorLine &edge1, const QDeclarativeAnchorLine &edge2, int offset1, int offset2, QDeclarativeAnchorLine::AnchorLine line, int &stretch);
+    bool calcStretch(const QDeclarativeAnchorLine &edge1, const QDeclarativeAnchorLine &edge2, qreal offset1, qreal offset2, QDeclarativeAnchorLine::AnchorLine line, qreal &stretch);
 
     void updateHorizontalAnchors();
     void updateVerticalAnchors();
